add fiat prices
